moved startServer to server.ts
