+++ packages/commons/src/context/context.ts
@@ -1,8 +1,6 @@
 import { constants } from "http2";
 import { randomUUID } from "crypto";
-import { Server } from "http";
 import {
-  ZodiosApp,
   ZodiosRouterContextRequestHandler,
   zodiosContext,
 } from "@zodios/express";
@@ -71,22 +69,3 @@ export const contextMiddleware =
 
     return next();
   };
-
-export function startServer(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  app: ZodiosApp<any, ExpressContext>,
-  config: {
-    host: string;
-    port: number;
-    keepAliveTimeout: number;
-  }
-): Server {
-  const server = app.listen(config.port, config.host, () => {
-    genericLogger.info(`Listening on ${config.host}:${config.port}`);
-  });
-
-  // eslint-disable-next-line functional/immutable-data
-  server.keepAliveTimeout = config.keepAliveTimeout;
-
-  return server;
-}
